* replace `short` with `ushort` for fields type of entity `ImageOcrBox` @ imagePipeline * replace `int` with `uint` for fields type of class `ClientRequesterTcs` @ crawler
* now allow using env-var and cli `args` to set or override config values in `appsettings.json`
* update request header user agent @ Strings.cs
@ shared
@ c#

+++ c#/crawler/src/Tieba/ClientRequesterTcs.cs
@@ -8,9 +8,9 @@ public class ClientRequesterTcs : WithLogTrace
     private readonly Timer _timer = new() {Enabled = true};
     private double _maxRps;
     private readonly Stopwatch _stopwatch = new();
-    private int _requestCounter;
+    private uint _requestCounter;
 
-    private int QueueLength => _queue.Count;
+    private uint QueueLength => (uint)_queue.Count;
     private float AverageRps => _requestCounter / (float)_stopwatch.Elapsed.TotalSeconds;
     private double MaxRps
     {
